the end

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ndesprez <ndesprez@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/03 14:07:06 by ndesprez          #+#    #+#             */
-/*   Updated: 2023/04/04 18:00:12 by ndesprez         ###   ########.fr       */
+/*   Updated: 2023/04/05 10:36:44 by ndesprez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
-int	ft_putnbr(int nb);
-int	ft_puthex(long long nb, char maj);
+int	ft_putnbr(long nb);
+int	ft_puthex(unsigned long long nb, char maj);
 int	ft_printf(const char *format, ...);
 
 #endif
